backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
tensorflow
scikit-learn
python-multipart
python-dotenv
openai
requests
voyageai
